+++ CHANGELOG.md
@@ -20,6 +20,8 @@
 - added cut corner image shape
 - added image stylizer
 - made offline lyrics sync for track instead of being global
+- added 10 band equalizer to android P and above
+- added custom eq preset to all versions
 
 ## 2.1.1
 - fixed all tracks sort order by date